Share HitReceptor between all object that uses it
