+++ osu.Game.Rulesets.Sentakki/Objects/Drawables/DrawableHold.cs
@@ -5,8 +5,6 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Input.Bindings;
-using osu.Framework.Input.Events;
 using osu.Game.Rulesets.Sentakki.Configuration;
 using osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces;
 using osu.Game.Rulesets.Sentakki.UI;
@@ -17,9 +15,6 @@ using osu.Game.Rulesets.Scoring;
 using osuTK;
 using osuTK.Graphics;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-
 namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
 {
     public class DrawableHold : DrawableSentakkiHitObject
@@ -243,89 +238,6 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
 
             HoldStartTime = Time.Current;
             isHitting.Value = true;
-        }
-
-        public class HitReceptor : CircularContainer, IKeyBindingHandler<SentakkiAction>
-        {
-            // IsHovered is used
-            public override bool HandlePositionalInput => true;
-
-            private SentakkiInputManager sentakkiActionInputManager;
-            internal SentakkiInputManager SentakkiActionInputManager => sentakkiActionInputManager ??= GetContainingInputManager() as SentakkiInputManager;
-
-            private List<SentakkiAction> actions = new List<SentakkiAction>();
-            public Func<bool> Hit;
-            public Action Release;
-
-            public float NoteAngle = -1;
-            public bool HoverAction()
-            {
-                if (!SentakkiActionInputManager.CurrentAngles.Contains(NoteAngle))
-                {
-                    SentakkiActionInputManager.CurrentAngles.Add(NoteAngle);
-                    if (SentakkiActionInputManager.PressedActions.Any(action => OnPressed(action)))
-                        actions.AddRange(SentakkiActionInputManager.PressedActions);
-                }
-                return false;
-            }
-            public HitReceptor()
-            {
-                RelativeSizeAxes = Axes.None;
-                Size = new Vector2(240);
-                Anchor = Anchor.Centre;
-                Origin = Anchor.Centre;
-                Add(new HoverReceptor(this));
-            }
-
-            public virtual bool OnPressed(SentakkiAction action)
-            {
-                switch (action)
-                {
-                    case SentakkiAction.Button1:
-                    case SentakkiAction.Button2:
-                        if (IsHovered && (Hit?.Invoke() ?? false))
-                        {
-                            actions.Add(action);
-                            return true;
-                        }
-                        break;
-                }
-                return false;
-            }
-
-            public void OnReleased(SentakkiAction action)
-            {
-                switch (action)
-                {
-                    case SentakkiAction.Button1:
-                    case SentakkiAction.Button2:
-                        if (actions.Contains(action))
-                            actions.Remove(action);
-                        if (!actions.Any())
-                            Release?.Invoke();
-                        break;
-                }
-            }
-            protected override void OnHoverLost(HoverLostEvent e)
-            {
-                SentakkiActionInputManager.CurrentAngles.Remove(NoteAngle);
-                actions.Clear();
-                Release?.Invoke();
-            }
-
-            public class HoverReceptor : CircularContainer
-            {
-                private readonly HitReceptor parent;
-                public HoverReceptor(HitReceptor parent)
-                {
-                    Size = new Vector2(150);
-                    Anchor = Anchor.Centre;
-                    Origin = Anchor.Centre;
-                    this.parent = parent;
-                }
-
-                protected override bool OnHover(HoverEvent e) => parent.HoverAction();
-            }
         }
     }
 }
