+++ osu.Game.Rulesets.Sentakki/Objects/Drawables/DrawableTap.cs
@@ -5,8 +5,6 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Input.Bindings;
-using osu.Framework.Input.Events;
 using osu.Game.Rulesets.Sentakki.Configuration;
 using osu.Game.Rulesets.Sentakki.Objects.Drawables.Pieces;
 using osu.Game.Rulesets.Objects.Drawables;
@@ -15,8 +13,6 @@ using osuTK;
 using osuTK.Graphics;
 using System;
 using System.Diagnostics;
-using System.Linq;
-
 namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
 {
     public class DrawableTap : DrawableSentakkiHitObject
@@ -61,7 +57,6 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
                         UpdateResult(true);
                         return true;
                     },
-                    RelativeSizeAxes = Axes.None,
                     Position = hitObject.EndPosition,
                     NoteAngle = HitObject.Angle
                 },
@@ -144,77 +139,6 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
 
                     break;
             }
-        }
-
-        public class HitReceptor : CircularContainer, IKeyBindingHandler<SentakkiAction>
-        {
-            // IsHovered is used
-            public override bool HandlePositionalInput => true;
-
-            public Func<bool> Hit;
-
-            public SentakkiAction? HitAction;
-
-            private SentakkiInputManager sentakkiActionInputManager;
-            internal SentakkiInputManager SentakkiActionInputManager => sentakkiActionInputManager ??= GetContainingInputManager() as SentakkiInputManager;
-
-            public float NoteAngle = -1;
-            public HitReceptor()
-            {
-                RelativeSizeAxes = Axes.None;
-                Size = new Vector2(240);
-                Anchor = Anchor.Centre;
-                Origin = Anchor.Centre;
-                Add(new HoverReceptor(this));
-            }
-
-            public bool HoverAction()
-            {
-                if (!SentakkiActionInputManager.CurrentAngles.Contains(NoteAngle))
-                {
-                    SentakkiActionInputManager.CurrentAngles.Add(NoteAngle);
-                    SentakkiActionInputManager?.PressedActions.Any(action => OnPressed(action));
-                }
-                return false;
-            }
-            public virtual bool OnPressed(SentakkiAction action)
-            {
-                switch (action)
-                {
-                    case SentakkiAction.Button1:
-                    case SentakkiAction.Button2:
-                        if (IsHovered && (Hit?.Invoke() ?? false))
-                        {
-                            HitAction = action;
-                            return true;
-                        }
-                        break;
-                }
-
-                return false;
-            }
-
-            public virtual void OnReleased(SentakkiAction action)
-            {
-            }
-            protected override void OnHoverLost(HoverLostEvent e)
-            {
-                SentakkiActionInputManager.CurrentAngles.Remove(NoteAngle);
-            }
-        }
-        public class HoverReceptor : CircularContainer
-        {
-            public override bool HandlePositionalInput => true;
-            private readonly HitReceptor parent;
-            public HoverReceptor(HitReceptor parent)
-            {
-                RelativeSizeAxes = Axes.None;
-                Size = new Vector2(150);
-                Anchor = Anchor.Centre;
-                Origin = Anchor.Centre;
-                this.parent = parent;
-            }
-            protected override bool OnHover(HoverEvent e) => parent.HoverAction();
         }
     }
 }
